Update and rename Jarvis.py to datalocker.py
